#Conflicts:
#	board/stm32f405-dk/src/boot/Makefile
#	board/stm32f405-dk/src/boot/board.c
#	board/stm32f405-dk/src/boot/boot.c
#	board/stm32f405-dk/src/include/board.h
#	board/stm32f405-dk/src/include/config.h
#	board/stm32f405-dk/src/sanity/Makefile
#	board/stm32f405-dk/src/sanity/prog.sh
#	ld/stm32f405.ld
#	ld/stm32f4xx.ld

+++ sdk/libc/vsnprintf.c
@@ -38,7 +38,7 @@ struct str_buf {
 	char * end;
 };
 
-static int sprintf_write(struct str_buf * str, const void * buf, int len)
+static int sprintf_write(struct str_buf * str, const void * buf, size_t len)
 {
 	char * src = (char *)buf;
 	char * dst = (char *)str->ptr;
@@ -59,7 +59,7 @@ static int sprintf_write(struct str_buf * str, const void * buf, int len)
 }
 
 static const struct fileop sprintf_fileop = {
-	.write = (int (*)(void *, const void *, int))sprintf_write
+	.write = (int (*)(void *, const void *, size_t))sprintf_write
 };
 
 /*
